Initial walk-in node support.
Walk-in nodes are now given the configuration of the skeleton
group applied atop the configuration of the default group.  Node
configurations will thus always be available when requested via
the internal API (and thus the REST API) for a valid commit.

This also switches cached node configurations to use the
lightweight format as in Wallaby classic.